neurotask/requirements.txt
openai>=1.0.0
SpeechRecognition>=3.10.0
pyttsx3>=2.90
python-dotenv>=1.0.0
ttkthemes>=3.2.2
pytest>=7.4.0
python-dateutil>=2.8.2
requests>=2.31.0
